edec_bot/requirements.txt
py-clob-client>=0.34.5
websockets>=12.0
httpx>=0.27.0
pycoingecko>=3.1.0
python-dotenv>=1.0.0
pyyaml>=6.0
python-telegram-bot>=21.0
openpyxl>=3.1.0
duckdb>=1.5.2
